requirements.txt
pyyaml==6.0
biopython==1.79
intarna==3.3.1
pandas==1.5.1